base/planner/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
langchain-core>=1.2.13,<2.0.0
langchain-openai>=1.1.10,<2.0.0
langgraph>=1.0.0,<2.0.0
pymilvus>=2.5.0
httpx>=0.28.0
prometheus-client>=0.21.0
pyyaml>=6.0
FlashRank>=0.2.10
rank-bm25>=0.2.2
kubernetes-asyncio>=31.0.0

# Test dependencies (not installed in production image)
# Install with: pip install -r requirements-test.txt
